+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed
   - glTF deduplication (incl. performance improvement)
+  - sqlite serialization when ignoring mesh, line, or point geometries
 
 ## [0.7.0] - 2019-10-24